% Generated by roxygen2: do not edit by hand
\name{Massloss-package}
\alias{Massloss-package}
\alias{Massloss}
\docType{package}
\title{
Mass Loss Analysis Package
}
\description{
A package for analyzing mass loss data. This package provides
    functions for processing, analyzing, and visualizing mass loss measurements.
}
\details{
\tabular{ll}{
Package: \tab Massloss\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2024-01-01\cr
License: \tab MIT\cr
}
}
\author{
Thiru Somasundaram <thiru.somasundaram@deakin.edu.au>
}
\references{
\url{https://github.com/DrThiruCSomasundaram/Massloss}
}
\keyword{package}
